requirements.txt
django==2.1.9
django-tinymce==2.8.0
stripe==2.26.0
gitpython==2.1.11
Pillow==6.0.0
pandas==0.24.2
folium==0.8.3
django-allauth==0.39.1
django-localflavor==2.2
bootstrap-admin==0.4.3
django-filter==2.1.0
django-mysql==3.2.0
django-qr-code
openpyxl